refactor(app): unmisfile three modules, hand ConfigColumn one object
Structure review, findings 17 and 19.

Three individual misfilings, each verified before moving rather than taken on
the review's word:

- `components/ThemeSwitch` -> `features/session/`. One consumer, `SessionMenu`,
  and it is a feature, not a generic widget.
- `collectGithubAuthFailures` -> `lib/github-failure.ts`, joining the
  `isGithubRateLimited` contract it is built on. It was the app SHELL's only
  runtime reach into a feature slice. `githubAuthFailure` and
  `githubPresetDisplayName` move with it because it calls them — leaving them
  behind would have meant a shared module importing from a feature, which is
  banned, and rightly.
- `AuthFailureBanner`'s own comment said the naming crossed the layer line "as
  a sentence's worth of data" precisely because the banner may not reach into
  the presets feature. With the walk in `lib/`, that contortion is just the
  normal downward import.

Finding 19: `ConfigColumn` took 40 props, twelve of which were one hook's
return — unpacked in App, re-listed in the props interface, re-destructured in
the body, then re-assembled for `RepoLoadOverlay`. That is exactly the
accretion roadmap 086 diagnosed for the results half and fixed with the
run-view context; the config half never got the same treatment. It takes
`repoLoad` whole now: twelve props become three.

The two that stay separate are the reference field (`repo`/`onRepoChange`),
because App owns that string — the inherited-config layer derives its probe
target from it, which `useRepoLoad`'s own header explains. The prop doc now
says so, so it does not look like an oversight.

Nothing is memoised on these props, so there was no render-count contract to
break. 968 tests green, lint and typecheck clean.
