+++ packages/app/src/features/presets/PresetDetail.tsx
@@ -9,13 +9,8 @@ import { PresetName } from "@/components/PresetName";
 import { findPollutedPath } from "@/lib/input-schemas";
 import type { NodeDescriptionFacts } from "@/lib/tree-descriptions";
 import { NodeDescriptionLines } from "./NodeDescriptions";
-import {
-  githubAuthFailure,
-  type InjectionKeyFn,
-  nodeInjectionKey,
-  type ParseFn,
-  STATE_LABELS,
-} from "./tree-shared";
+import { githubAuthFailure } from "@/lib/github-failure";
+import { type InjectionKeyFn, nodeInjectionKey, type ParseFn, STATE_LABELS } from "./tree-shared";
 import { useEngineHelpers } from "./use-engine-helpers";
 import { pluralWord } from "@/lib/format";
 import { errorMessage } from "@/lib/errors";
